Migrate disruption to use standard component
